开始制作测试计划

+++ backend/app/models/order.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, Float, ForeignKey, DateTime, func
+from sqlalchemy import Column, Integer, Float, ForeignKey, DateTime, func, Numeric
 from sqlalchemy.orm import relationship
 from sqlalchemy.dialects.mssql import NVARCHAR
 
@@ -10,7 +10,7 @@ class Order(Base):
     id = Column(Integer, primary_key=True, index=True)
     order_sn = Column(NVARCHAR(64), unique=True, index=True, nullable=False)
     user_id = Column(Integer)  # Assuming you have a User model and will add a ForeignKey
-    total_amount = Column(Float, nullable=False)
+    total_amount = Column(Numeric(10, 2), nullable=False)
     status = Column(NVARCHAR(50), default='PENDING')
     shipping_address = Column(NVARCHAR(255), nullable=False)
     shipping_contact = Column(NVARCHAR(255))
@@ -26,7 +26,7 @@ class OrderItem(Base):
     order_id = Column(Integer, ForeignKey("orders.id"))
     product_id = Column(Integer, ForeignKey("products.id"))
     quantity = Column(Integer, nullable=False)
-    price = Column(Float, nullable=False)
+    price = Column(Numeric(10, 2), nullable=False)
 
     order = relationship("Order", back_populates="items")
     product = relationship("Product")
